requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.7
et_xmlfile==2.0.0
idna==3.11
openpyxl==3.1.5
pillow==12.2.0
python-dotenv==1.2.2
reportlab==4.4.10
requests==2.33.1
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
